Move password hashing and verifying functions to class User
