+++ Technolab.OnlineLibrary.Web/Controllers/AuthController.cs
@@ -5,8 +5,6 @@ using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Technolab.OnlineLibrary.Web.Models;
 using Technolab.OnlineLibrary.Web.ViewModels;
-using System.Security.Cryptography;
-using System.Text;
 
 namespace Technolab.OnlineLibrary.Web.Controllers
 {
@@ -31,7 +29,7 @@ namespace Technolab.OnlineLibrary.Web.Controllers
             using var context = ContextFactory.Create();
 
             var user = context.Users
-                .Where(x => x.Username == model.Username && VerifyPassword(model.Password, x.PasswordHash))
+                .Where(x => x.Username == model.Username && x.VerifyPassword(model.Password))
                 .SingleOrDefault();
             if (user == null)
             {
@@ -70,42 +68,5 @@ namespace Technolab.OnlineLibrary.Web.Controllers
         }
 
         private ILibraryDbContextFactory ContextFactory { get; }
-
-        private bool VerifyPassword(string enteredPassword, string correctHashWithParams)
-        {
-            string[] split = correctHashWithParams.Split(":");
-
-            string correctHash = split[0];
-            byte[] salt = Convert.FromBase64String(split[1]);
-            int iterations = Convert.ToInt32(split[2]);
-            HashAlgorithmName hashAlgorithm = new HashAlgorithmName(split[3]);
-            int keySize = Convert.ToInt32(split[4]);
-
-            var enteredPasswordHash = Rfc2898DeriveBytes.Pbkdf2(enteredPassword, salt, iterations, hashAlgorithm, keySize);
-            return CryptographicOperations.FixedTimeEquals(enteredPasswordHash, Convert.FromHexString(correctHash));
-        }
-
-        private string GenerateHash(string password)
-        {
-            int keySize = 32;
-            byte[] salt = RandomNumberGenerator.GetBytes(keySize);
-            int iterations = 200000;
-            HashAlgorithmName hashAlgorithm = HashAlgorithmName.SHA512;
-
-            var hash = Rfc2898DeriveBytes.Pbkdf2(
-                Encoding.UTF8.GetBytes(password),
-                salt,
-                iterations,
-                hashAlgorithm,
-                keySize);
-
-            string hashStr = Convert.ToHexString(hash) + ":" +
-                             Convert.ToBase64String(salt) + ":" +
-                             iterations + ":" +
-                             hashAlgorithm.Name + ":" + 
-                             keySize;
-
-            return hashStr;
-        }
     }
 }
